Integration of DIJ into PyImageJ
Include an extra parameter in the headless mode MAcro arguments (models_dir) that can be used to provide the models directory we want DIJ to look at.
This is done to be fully compatible with PyImageJ. Before this, DIJ PyImageJ did not know where was the models folder. Prior to this change DIJ looked for the directory from where IJ/Fiji was launched, but as with PyIMageJ, IJ/Fiji can be launched from anywhere, some errors arised as the directory was not always IJ/Fiji directories